+++ src/main/java/QUT/CAB302/fortunecookie/Model/BackNavigable.java
@@ -1,4 +1,4 @@
-package QUT.CAB302.fortunecookie;
+package QUT.CAB302.fortunecookie.Model;
 
 public interface BackNavigable {
     void goBack();
