+++ src/main/java/QUT/CAB302/fortunecookie/Model/User.java
@@ -1,4 +1,4 @@
-package QUT.CAB302.fortunecookie;
+package QUT.CAB302.fortunecookie.Model;
 
 // A class to represent user's credentials
 public class User {
