+++ src/main/java/QUT/CAB302/fortunecookie/Model/UserDAODatabase.java
@@ -1,4 +1,4 @@
-package QUT.CAB302.fortunecookie;
+package QUT.CAB302.fortunecookie.Model;
 
 import java.sql.*;
 import org.mindrot.jbcrypt.BCrypt;
