+++ src/main/resources/QUT/CAB302/fortunecookie/FortuneCookie.png
