--- src/main/resources/QUT/CAB302/fortunecookie/LegacyFiles/hello-view.fxml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.layout.HBox?>
-<VBox alignment="TOP_LEFT" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="QUT.CAB302.fortunecookie.LegacyFiles.HelloController">
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
-    </padding>
-
-    <Label text="Welcome to the Address Book Application.
-  You must agree to the following terms and conditions before you can use the application."
-           wrapText="true"/>
-    <TextArea fx:id="termsAndConditions"
-              wrapText="true"
-              editable="false"
-    />
-    <CheckBox fx:id="agreeCheckBox"
-              text="I agree to the terms and conditions." onAction="#onAgreeCheckBoxClick" />
-
-    <HBox alignment="BOTTOM_RIGHT" spacing="10.0">
-        <Button text="Cancel" onAction="#onCancelButtonClick"/>
-        <Button fx:id="nextButton" text="Next" onAction="#onNextButtonClick" disable="true"/>
-    </HBox>
-    <Label fx:id="welcomeText"/>
-    <Button text="Hello!" onAction="#onHelloButtonClick"/>
-</VBox>
