--- src/main/resources/QUT/CAB302/fortunecookie/LegacyFiles/main-view.fxml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<VBox alignment="TOP_LEFT" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="QUT.CAB302.fortunecookie.LegacyFiles.MainController"
->
-    <padding>
-        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
-    </padding>
-    <Label text="Select a contact to view or edit." />
-    <HBox VBox.vgrow="ALWAYS" spacing="20.0">
-        <VBox spacing="10.0">
-            <ListView fx:id="contactsListView"  />
-            <Button text="New" maxWidth="Infinity"/>
-        </VBox>
-        <VBox spacing="10.0" prefWidth="400" fx:id="contactContainer">
-            <Label text="Enter the contact's details below." />
-            <GridPane hgap="5.0" vgap="5.0" >
-                <columnConstraints>
-                    <ColumnConstraints minWidth="70" />
-                    <ColumnConstraints hgrow="ALWAYS"/>
-                </columnConstraints>
-                <Label text="First Name:" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-                <Label text="Last Name:" GridPane.columnIndex="0" GridPane.rowIndex="1" />
-                <Label text="Email:" GridPane.columnIndex="0" GridPane.rowIndex="2" />
-                <Label text="Phone:" GridPane.columnIndex="0" GridPane.rowIndex="3" />
-                <TextField fx:id="firstNameTextField" GridPane.columnIndex="1" GridPane.rowIndex="0" maxWidth="Infinity"/>
-                <TextField fx:id="lastNameTextField" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-                <TextField fx:id="emailTextField" GridPane.columnIndex="1" GridPane.rowIndex="2" />
-                <TextField fx:id="phoneTextField" GridPane.columnIndex="1" GridPane.rowIndex="3" />
-            </GridPane>
-            <HBox spacing="10.0" alignment="CENTER">
-                <Button text="Confirm" onAction="#onEditConfirm"/>
-                <Button text="Cancel" onAction="#onCancel"/>
-                <Button text="Delete" onAction="#onDelete"/>
-                <Button text="New" onAction="#onAdd" maxWidth="Infinity"/>
-            </HBox>
-
-        </VBox>
-    </HBox>
-</VBox>
