+++ src/main/resources/QUT/CAB302/fortunecookie/Login.fxml
@@ -13,7 +13,7 @@
 
 <BorderPane xmlns="http://javafx.com/javafx/23.0.1"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="QUT.CAB302.fortunecookie.LoginController"
+            fx:controller="QUT.CAB302.fortunecookie.Controller.LoginController"
             style="-fx-font-family: 'Aptos'; -fx-background-color: #FFF9F2;">
 
     <!-- BIG header: enlarged cookie + elegant title -->
