+++ src/main/resources/QUT/CAB302/fortunecookie/Settings.fxml
@@ -5,7 +5,7 @@
 <?import java.lang.String?>
 <?import javafx.collections.FXCollections?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #FFF9F2;" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="QUT.CAB302.fortunecookie.AccountSettingsController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #FFF9F2;" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="QUT.CAB302.fortunecookie.Controller.AccountSettingsController">
     <children>
         <VBox layoutX="20" layoutY="20" prefWidth="400" spacing="10">
