+++ src/main/resources/QUT/CAB302/fortunecookie/SettingsIcon.png
