+++ src/main/resources/QUT/CAB302/fortunecookie/StudyMode.fxml
@@ -19,7 +19,7 @@
 
 <BorderPane xmlns="http://javafx.com/javafx/23.0.1"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="QUT.CAB302.fortunecookie.StudyModeController"
+            fx:controller="QUT.CAB302.fortunecookie.Controller.StudyModeController"
             style="-fx-background-color: #FFF9F2;">
 
     <!-- Top bar: Study Mode header with Back and LOG OUT -->
